refactor(ai): move knowledge base search from plugin-ai to plugin-ai-knowledge-base
