+++ packages/plugins/@nocobase/plugin-ai/src/server/features/vector-store-provider.ts
@@ -7,7 +7,7 @@
  * For more information, please refer to: https://www.nocobase.com/agreement.
  */
 
-import { VectorStoreProp } from '../types';
+import { DocumentSegmentedWithScore, VectorStoreProp } from '../types';
 
 export interface VectorStoreProviderFeature {
   register(vsp: VectorStoreProvider): void;
@@ -29,14 +29,4 @@ export type VectorStoreSearchOptions = {
   topK?: number;
   score?: string;
   filter?: any;
-};
-
-export type DocumentSegmented = {
-  content: string;
-  metadata: Record<string, any>;
-  id?: string;
-};
-
-export type DocumentSegmentedWithScore = DocumentSegmented & {
-  score: number;
 };
