+++ packages/plugins/@nocobase/plugin-ai/src/server/types/knowledge-base.type.ts
@@ -37,3 +37,15 @@ export type KnowledgeBaseGroup = {
   knowledgeBaseType: KnowledgeBaseType;
   knowledgeBaseList: KnowledgeBase[];
 };
+
+export type DocumentSegmented = {
+  content: string;
+  metadata: Record<string, any>;
+  id?: string;
+};
+
+export type DocumentSegmentedWithScore = DocumentSegmented & {
+  score: number;
+};
+
+export type SearchOptions = { knowledgeBaseKeys: string[]; query: string; topK?: number; score?: string };
